requirements.txt
python-telegram-bot>=20.6,<22
python-dotenv>=1.0,<2
